split test between path and exceptions
