+++ packages/contracts/contracts/RainbowManager.sol
@@ -21,7 +21,10 @@ contract RainbowManager {
     uint8 g,
     uint8 b
   ) external {
-    // require(r > 5 && r < 250 && g > 5 && g < 250 && b > 5 && b < 250, "Target color is too close to a base color");
+    require(
+      r > 5 && r < 250 && g > 5 && g < 250 && b > 5 && b < 250,
+      "Target color is too close to a base color"
+    );
     game.claimVictory(msg.sender);
     address managerAddress = address(this);
     game = new RainbowToken(managerAddress, r, g, b);
