+++ packages/contracts/test/rainbow-token.js
@@ -1,3 +1,4 @@
+const RainbowToken = artifacts.require("RainbowToken");
 const {
   getBalances,
   expectDecrease,
@@ -10,7 +11,7 @@ const {
   addPlayersToGame,
 } = require("./utils");
 
-contract("RainbowToken", (accounts) => {
+contract("RainbowToken Game Path", (accounts) => {
   it("should allow various addresses to join the game and blend together", async () => {
     const [deployer, ...playerAddresses] = accounts;
 
@@ -52,7 +53,12 @@ contract("RainbowToken", (accounts) => {
       customSetup,
     });
 
-    // Balances should have decreased by 0.1 ether for each of the player
+    /**
+     * Players have joined the game
+     * Balances should have decreased by 0.1 ether for each player
+     * Balance of the contract should have increased by 0.1 * number of players
+     * Player addresses should be in the player list
+     */
 
     const updatedPlayerBalances = await getBalances(playerAddresses);
     playerAddresses.forEach((_, i) => {
@@ -64,7 +70,6 @@ contract("RainbowToken", (accounts) => {
       );
     });
 
-    // Contract balance should have increased
     let [contractBalance] = await getBalances([instance.address]);
     expectIncrease(
       0,
@@ -81,13 +86,22 @@ contract("RainbowToken", (accounts) => {
       );
     });
 
-    // Other player increases its blending price
+    /**
+     * Second player increases its blending price
+     */
     await instance.setBlendingPrice(web3.utils.toWei("1"), {
       from: secondPlayer.address,
       gasPrice: 0,
     });
 
-    // First player blends with another player
+    /**
+     * First player blends with second player
+     * Color of the first player should be updated
+     * Color of the second player should not be updated
+     * Balance of the first player should have decreased by 1 ether
+     * Balance of the second player should have increased by 0.5 ether
+     * Balance of the contract should have increased by 0.5 ether
+     */
 
     await instance.blend(
       secondPlayer.address,
@@ -96,8 +110,6 @@ contract("RainbowToken", (accounts) => {
       secondPlayer.color.b,
       { from: firstPlayer.address, gasPrice: 0, value: web3.utils.toWei("1") }
     );
-
-    // Color should be updated for the first player only
 
     const [updatedFirstPlayer, updatedSecondPlayer] = await instance.getPlayers(
       [firstPlayer.address, secondPlayer.address]
@@ -133,8 +145,6 @@ contract("RainbowToken", (accounts) => {
       "Other player B component has been updated after blend with first player"
     );
 
-    // Balances should be updated
-
     const [updatedFirstPlayerBalance, updatedSecondPlayerBalance] =
       await getBalances([firstPlayer.address, secondPlayer.address]);
     expectDecrease(
@@ -159,15 +169,18 @@ contract("RainbowToken", (accounts) => {
       "Balance of the contract is not correct after blend between the two players"
     );
 
-    // First player does a self blend
+    /**
+     * First player self blends
+     * Color of the first player should be updated
+     * Balance of the first player should have decreased by 0.5 ether
+     * Balance of the contract should have increased by 0.5 ether
+     */
 
     await instance.selfBlend({
       from: firstPlayer.address,
       gasPrice: 0,
       value: web3.utils.toWei("0.5"),
     });
-
-    // Color should be updated
 
     const firstPlayerFinalState = await instance.getPlayer(firstPlayer.address);
     expect(Number(firstPlayerFinalState.color.r)).to.equal(
@@ -195,8 +208,6 @@ contract("RainbowToken", (accounts) => {
       "Player B component has been wrongly updated after self blend"
     );
 
-    // Balances should be updated
-
     const [finalFirstPlayerBalance] = await getBalances([firstPlayer.address]);
     expectDecrease(
       updatedFirstPlayerBalance,
@@ -214,7 +225,13 @@ contract("RainbowToken", (accounts) => {
       "Balance of the contract is not correct after self blend"
     );
 
-    // Winner claims victory
+    /**
+     * Winner claims victory
+     * Contract balance should be 0
+     * Balance of the winner should have increased by the amount of ether in the contract
+     * Game should be over
+     * New RainbowToken should be created with no player and new target color
+     */
 
     const [winner, matchingPlayer] = winnerPair;
     await instance.blend(
@@ -232,7 +249,7 @@ contract("RainbowToken", (accounts) => {
     [contractBalance] = await getBalances([instance.address]);
 
     const [beforeWinWinnerBalance] = await getBalances([winner.address]);
-    await managerInstance.claimVictory(0, 0, 0, {
+    await managerInstance.claimVictory(100, 100, 100, {
       from: winner.address,
       gasPrice: 0,
     });
@@ -252,471 +269,16 @@ contract("RainbowToken", (accounts) => {
     const gameOver = await instance.gameOver();
 
     expect(gameOver).to.equal(true);
-  });
 
-  describe("Exceptions", () => {
-    const gameOverSetup =
-      (playerAddresses) =>
-      async ({ instance, managerInstance }) => {
-        const players = await addPlayersToGame({ instance, playerAddresses });
-        const winnerPair = findWinnerPair({
-          players,
-          targetColor: { r: 127, g: 127, b: 127 },
-        });
+    const newContractAddress = await managerInstance.getGameAddress();
+    const newInstance = await RainbowToken.at(newContractAddress);
 
-        const [winner, matchingPlayer] = winnerPair;
-        await instance.blend(
-          matchingPlayer.address,
-          matchingPlayer.color.r,
-          matchingPlayer.color.g,
-          matchingPlayer.color.b,
-          {
-            from: winner.address,
-            value: web3.utils.toWei("0.1"),
-            gasPrice: 0,
-          }
-        );
+    const targetColor = await newInstance.getTargetColor();
+    expect(targetColor.r).to.equal("100");
+    expect(targetColor.g).to.equal("100");
+    expect(targetColor.b).to.equal("100");
 
-        await managerInstance.claimVictory(127, 127, 127, {
-          from: winner.address,
-          gasPrice: 0,
-        });
-
-        return winnerPair;
-      };
-    describe("joining the game", () => {
-      it("should revert if not enough value is sent to join the game", async () => {
-        const [deployer, player] = accounts;
-
-        const { instance } = await testSetup({ deployer });
-
-        await expectRevert(
-          instance.joinGame({
-            from: player,
-            value: web3.utils.toWei("0.09"),
-            gasPrice: 0,
-          }),
-          "Not enough sent Ethers to join the game"
-        );
-      });
-
-      it("should revert if sender is already a player", async () => {
-        const [deployer, account] = accounts;
-
-        const { instance } = await testSetup({ deployer });
-
-        await instance.joinGame({
-          from: account,
-          value: web3.utils.toWei("0.1"),
-          gasPrice: 0,
-        });
-
-        await expectRevert(
-          instance.joinGame({
-            from: account,
-            value: web3.utils.toWei("0.1"),
-            gasPrice: 0,
-          }),
-          "Sender is already a player"
-        );
-      });
-
-      it("should revert if game is over", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const { instance } = await testSetup({
-          deployer,
-          customSetup: gameOverSetup(playerAddresses),
-        });
-
-        await expectRevert(
-          instance.joinGame({
-            from: deployer,
-            value: web3.utils.toWei("0.1"),
-            gasPrice: 0,
-          }),
-          "The game is over"
-        );
-      });
-    });
-
-    describe("self blend", () => {
-      it("should revert if not enough value is sent", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const {
-          instance,
-          custom: { firstPlayer },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses,
-            });
-            const firstPlayer = findRandomPlayer({ players });
-            return { firstPlayer };
-          },
-        });
-
-        await expectRevert(
-          instance.selfBlend({
-            from: firstPlayer.address,
-            value: web3.utils.toWei("0.49"),
-            gasPrice: 0,
-          }),
-          "Not enough sent Ethers to self blend"
-        );
-      });
-
-      it("should revert if sender is not a player", async () => {
-        const [deployer, account] = accounts;
-
-        const { instance } = await testSetup({ deployer });
-
-        await expectRevert(
-          instance.selfBlend({
-            from: account,
-            value: web3.utils.toWei("0.5"),
-            gasPrice: 0,
-          }),
-          "Sender is not a player"
-        );
-      });
-
-      it("should revert if game is over", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const {
-          instance,
-          custom: [, firstPlayer],
-        } = await testSetup({
-          deployer,
-          customSetup: gameOverSetup(playerAddresses),
-        });
-
-        await expectRevert(
-          instance.selfBlend({
-            from: firstPlayer.address,
-            value: web3.utils.toWei("0.5"),
-            gasPrice: 0,
-          }),
-          "The game is over"
-        );
-      });
-    });
-
-    describe("blend", () => {
-      it("should revert if not enough value is sent", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const {
-          instance,
-          custom: { firstPlayer, secondPlayer },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses,
-            });
-            const firstPlayer = findRandomPlayer({ players });
-            const secondPlayer = findRandomPlayer({
-              players,
-              excludedPlayers: [firstPlayer],
-            });
-            return {
-              firstPlayer,
-              secondPlayer,
-            };
-          },
-        });
-
-        await expectRevert(
-          instance.blend(
-            secondPlayer.address,
-            secondPlayer.color.r,
-            secondPlayer.color.g,
-            secondPlayer.color.b,
-            {
-              from: firstPlayer.address,
-              value: web3.utils.toWei("0.09"),
-              gasPrice: 0,
-            }
-          ),
-          "Not enough sent Ethers to blend"
-        );
-      });
-
-      it("should revert if sender is not a player", async () => {
-        const [deployer, account] = accounts;
-
-        const {
-          instance,
-          custom: { firstPlayer },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses: [account],
-            });
-            const firstPlayer = findRandomPlayer({ players });
-            return {
-              firstPlayer,
-            };
-          },
-        });
-
-        await expectRevert(
-          instance.blend(
-            firstPlayer.address,
-            firstPlayer.color.r,
-            firstPlayer.color.g,
-            firstPlayer.color.b,
-            { from: deployer, value: web3.utils.toWei("0.1"), gasPrice: 0 }
-          ),
-          "Sender is not a player"
-        );
-      });
-
-      it("should revert if target is not a player", async () => {
-        const [deployer, account] = accounts;
-
-        const {
-          instance,
-          custom: { firstPlayer },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses: [account],
-            });
-            const firstPlayer = findRandomPlayer({ players });
-            return { firstPlayer };
-          },
-        });
-
-        await expectRevert(
-          instance.blend(deployer, 0, 0, 0, {
-            from: firstPlayer.address,
-            value: web3.utils.toWei("0.1"),
-            gasPrice: 0,
-          }),
-          "Target address is not a player"
-        );
-      });
-
-      it("should revert if input color is not the one of the player", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const {
-          instance,
-          custom: { firstPlayer, secondPlayer },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses,
-            });
-            const firstPlayer = findRandomPlayer({ players });
-            const secondPlayer = findAnotherColorPlayer({
-              players,
-              color: firstPlayer.color,
-              excludedPlayers: [firstPlayer],
-            });
-            return { firstPlayer, secondPlayer };
-          },
-        });
-
-        await expectRevert(
-          instance.blend(
-            secondPlayer.address,
-            firstPlayer.color.r,
-            firstPlayer.color.g,
-            firstPlayer.color.b,
-            {
-              from: firstPlayer.address,
-              value: web3.utils.toWei("0.1"),
-              gasPrice: 0,
-            }
-          ),
-          "Color of the other player has changed, blend reverted"
-        );
-      });
-
-      it("should revert if game is over", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const {
-          instance,
-          custom: [, otherPlayer],
-        } = await testSetup({
-          deployer,
-          customSetup: gameOverSetup(playerAddresses),
-        });
-
-        await expectRevert(
-          instance.blend(
-            otherPlayer.address,
-            otherPlayer.color.r,
-            otherPlayer.color.g,
-            otherPlayer.color.b,
-            { from: deployer, value: web3.utils.toWei("0.1"), gasPrice: 0 }
-          ),
-          "The game is over"
-        );
-      });
-    });
-
-    describe("setBlendingPrice", () => {
-      it("should revert if sender is not a player", async () => {
-        const [deployer, account] = accounts;
-
-        const { instance } = await testSetup({ deployer });
-
-        await expectRevert(
-          instance.setBlendingPrice(web3.utils.toWei("0.5"), {
-            from: account,
-            gasPrice: 0,
-          }),
-          "Sender is not a player"
-        );
-      });
-
-      it("should revert if new blending price is 0", async () => {
-        const [deployer, account] = accounts;
-
-        const {
-          instance,
-          custom: { firstPlayer },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses: [account],
-            });
-            const firstPlayer = findRandomPlayer({ players });
-            return { firstPlayer };
-          },
-        });
-
-        await expectRevert(
-          instance.setBlendingPrice(web3.utils.toWei("0"), {
-            from: firstPlayer.address,
-            gasPrice: 0,
-          }),
-          "Blending price must be higher than 0"
-        );
-      });
-
-      it("should revert if game is over", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-
-        const {
-          instance,
-          custom: [, otherPlayer],
-        } = await testSetup({
-          deployer,
-          customSetup: gameOverSetup(playerAddresses),
-        });
-
-        await expectRevert(
-          instance.setBlendingPrice(web3.utils.toWei("0.5"), {
-            from: otherPlayer.address,
-            gasPrice: 0,
-          }),
-          "The game is over"
-        );
-      });
-    });
-
-    describe("claimVictory", () => {
-      it("should revert if the sender is not a player", async () => {
-        const [deployer] = accounts;
-        const { managerInstance } = await testSetup({ deployer });
-
-        await expectRevert(
-          managerInstance.claimVictory(127, 127, 127, {
-            from: deployer,
-            gasPrice: 0,
-          }),
-          "Claimed winner address is not a player"
-        );
-      });
-
-      it("should revert if the game is over", async () => {
-        const [deployer, ...playerAddresses] = accounts;
-        const {
-          managerInstance,
-          custom: { firstWinner, secondWinner },
-        } = await testSetup({
-          deployer,
-          customSetup: async ({ instance }) => {
-            const players = await addPlayersToGame({
-              instance,
-              playerAddresses,
-            });
-
-            const firstWinnerPair = findWinnerPair({
-              players,
-              targetColor: { r: 127, g: 127, b: 127 },
-            });
-
-            const secondWinnerPair = findWinnerPair({
-              players,
-              targetColor: { r: 127, g: 127, b: 127 },
-              excludedPlayers: [firstWinnerPair[0]],
-            });
-
-            const [firstWinner, firstMatchingPlayer] = firstWinnerPair;
-            await instance.blend(
-              firstMatchingPlayer.address,
-              firstMatchingPlayer.color.r,
-              firstMatchingPlayer.color.g,
-              firstMatchingPlayer.color.b,
-              {
-                from: firstWinner.address,
-                value: web3.utils.toWei("0.1"),
-                gasPrice: 0,
-              }
-            );
-
-            const [secondWinner, secondMatchingPlayer] = secondWinnerPair;
-            await instance.blend(
-              secondMatchingPlayer.address,
-              secondMatchingPlayer.color.r,
-              secondMatchingPlayer.color.g,
-              secondMatchingPlayer.color.b,
-              {
-                from: secondWinner.address,
-                value: web3.utils.toWei("0.1"),
-                gasPrice: 0,
-              }
-            );
-
-            return {
-              firstWinner,
-              secondWinner,
-            };
-          },
-        });
-
-        await managerInstance.claimVictory(0, 0, 0, {
-          from: firstWinner.address,
-          gasPrice: 0,
-        }),
-          await expectRevert(
-            managerInstance.claimVictory(0, 0, 0, {
-              from: secondWinner.address,
-              gasPrice: 0,
-            }),
-            "Claimed winner address is not a player"
-          );
-      });
-    });
+    const newPlayerList = await newInstance.getPlayerList();
+    expect(newPlayerList.length).to.equal(0);
   });
 });
